refactor: TopBar 컴포넌트 분리
